// Desktop/UNEXT_LMS_PORTAL/src/app/enums/storageKey.ts
export enum StorageKey {
    LOGIN_CONFIG = 'LOGIN_CONFIG',
    USER_DETAILS = 'USER_DETAILS',
    BRANDING_CONFIG = 'BRANDING_CONFIG',
    JWT_TOKEN_DATA = 'JWT_TOKEN_DATA',
    COMMON_CONFIG = 'COMMON_CONFIG',
    DOMAIN_CONFIG = 'DOMAIN_CONFIG',
    ELEMENT_DETAIL = 'ELEMENT_DETAIL',
    USER_CURRENT_VIEW = 'USER_CURRENT_VIEW',
    DOC_VERSION = 'DOC_VERSION',
    COURSE_JSON = 'COURSE_JSON',
    CONTENT_PANE_SELECTED_QUESTION = 'CONTENT_PANE_SELECTED_QUESTION',
    ATTACHED_RUBRIC = 'ATTACHED_RUBRIC',
    ELEMENT_ID = 'ELEMENT_ID',
    DISCUSSION_FORUM_ACTIVE_TAB = 'DISCUSSION_FORUM_ACTIVE_TAB',
    COURSE_LIST_SELECTION = 'COURSE_LIST_SELECTION',
    SIDEBAR_ACTIVE_COMPONENT_ROUTE = 'SIDEBAR_ACTIVE_COMPONENT_ROUTE',
    SELF_TASK_OPERATION = 'SELF_TASK_OPERATION',
    CURRENT_DEVICE = 'CURRENT_DEVICE',
    }